feat: RSS URL 수정(뱅크샐러드/카카오뱅크/Stripe 등), 로고 URL 추가, MediaRSS 썸네일 추출, RSS 소스 리서치 문서화

+++ backend/src/main/resources/db/schema.sql
@@ -35,7 +35,7 @@ EXECUTE FUNCTION tsvector_update_trigger(search_vector, 'pg_catalog.simple', tit
 
 CREATE TABLE IF NOT EXISTS crawl_logs (
     id             BIGSERIAL PRIMARY KEY,
-    company_id     INT REFERENCES companies (id),
+    company_id     BIGINT REFERENCES companies (id),
     status         VARCHAR(20),
     articles_added INT DEFAULT 0,
     error_message  TEXT,
